Rename BasePeerList to PeerList
